all CUDA versions compiling

+++ test/performance-regression/cpu-only/rodinia/particlefilter/ex_particle_OPENMP_seq.ref.c
@@ -418,7 +418,7 @@ void particleFilter(int * I, int IszX, int IszY, int Nfr, int * seed, int Nparti
 			for(y = 0; y < countOnes; y++){
 				indX = roundDouble(arrayX[x]) + objxy[y*2 + 1];
 				indY = roundDouble(arrayY[x]) + objxy[y*2];
-				ind[x*countOnes + y] = fabs(indX*IszY*Nfr + indY*Nfr + k);
+				ind[x*countOnes + y] = fabs((double)(indX*IszY*Nfr + indY*Nfr + k));
 				if(ind[x*countOnes + y] >= max_size)
 					ind[x*countOnes + y] = 0;
 			}
